// src/api/client/api.ts
export interface ClientsApiGetClientsRequest {
  /**
   * The starting index of the invoices.
   * @type {number}
   * @memberof ClientsApiGetClients
   */
  readonly from?: number;

  /**
   * The number of invoices to return.
   * @type {number}
   * @memberof ClientsApiGetClients
   */
  readonly size?: number;

  /**
   * Filter criteria for invoices.
   * @type {string}
   * @memberof ClientsApiGetClients
   */
  readonly filter?: string;

  /**
   * Field to order clients by.
   * @type {'createdAt'}
   * @memberof ClientsApiGetClients
   */
  readonly orderBy?: GetClientsOrderByEnum;

  /**
   * Sorting direction.
   * @type {'ASC' | 'DESC'}
   * @memberof ClientsApiGetClients
   */
  readonly direction?: GetClientsDirectionEnum;

  /**
   * Filter clients by their status.
   * @type {'ACTIVE' | 'INACTIVE'}
   * @memberof ClientsApiGetClients
   */
  readonly status?: GetClientsStatusEnum;
}
// ...
export const GetClientsOrderByEnum = {
  CreatedAt: 'createdAt',
} as const;
export type GetClientsOrderByEnum = (typeof GetClientsOrderByEnum)[keyof typeof GetClientsOrderByEnum];
/**
 * @export
 */
export const GetClientsDirectionEnum = {
  Asc: 'ASC',
  Desc: 'DESC',
} as const;
export type GetClientsDirectionEnum = (typeof GetClientsDirectionEnum)[keyof typeof GetClientsDirectionEnum];
/**
 * @export
 */
export const GetClientsStatusEnum = {
  Active: 'ACTIVE',
  Inactive: 'INACTIVE',
} as const;
export type GetClientsStatusEnum = (typeof GetClientsStatusEnum)[keyof typeof GetClientsStatusEnum];