// src/api/documents/api.ts
export interface TemplatesApiGetTemplatesRequest {
    /**
     * Pagination start index
     * @type {number}
     * @memberof TemplatesApiGetTemplates
     */
    readonly from?: number

    /**
     * Number of items to return
     * @type {number}
     * @memberof TemplatesApiGetTemplates
     */
    readonly size?: number

    /**
     * Filter condition
     * @type {string}
     * @memberof TemplatesApiGetTemplates
     */
    readonly filter?: string

    /**
     * Field to order by
     * @type {string}
     * @memberof TemplatesApiGetTemplates
     */
    readonly orderBy?: string

    /**
     * 
     * @type {'ASC' | 'DESC'}
     * @memberof TemplatesApiGetTemplates
     */
    readonly direction?: GetTemplatesDirectionEnum
}
// ...
export const GetTemplatesDirectionEnum = {
    Asc: 'ASC',
    Desc: 'DESC'
} as const;
export type GetTemplatesDirectionEnum = typeof GetTemplatesDirectionEnum[keyof typeof GetTemplatesDirectionEnum];