// src/api/products/api.ts
export interface DefaultApiGetAllProductsRequest {
    /**
     * The starting index of the products.
     * @type {number}
     * @memberof DefaultApiGetAllProducts
     */
    readonly from?: number

    /**
     * The number of products to return.
     * @type {number}
     * @memberof DefaultApiGetAllProducts
     */
    readonly size?: number

    /**
     * Filter criteria for products.
     * @type {string}
     * @memberof DefaultApiGetAllProducts
     */
    readonly filter?: string

    /**
     * Field to order products by.
     * @type {'createdAt'}
     * @memberof DefaultApiGetAllProducts
     */
    readonly orderBy?: GetAllProductsOrderByEnum

    /**
     * Sorting direction.
     * @type {'ASC' | 'DESC'}
     * @memberof DefaultApiGetAllProducts
     */
    readonly direction?: GetAllProductsDirectionEnum

    /**
     * Filter product by their status.
     * @type {'ACTIVE' | 'INACTIVE'}
     * @memberof DefaultApiGetAllProducts
     */
    readonly status?: GetAllProductsStatusEnum
}
// ...
export const GetAllProductsOrderByEnum = {
    CreatedAt: 'createdAt'
} as const;
export type GetAllProductsOrderByEnum = typeof GetAllProductsOrderByEnum[keyof typeof GetAllProductsOrderByEnum];
/**
 * @export
 */
export const GetAllProductsDirectionEnum = {
    Asc: 'ASC',
    Desc: 'DESC'
} as const;
export type GetAllProductsDirectionEnum = typeof GetAllProductsDirectionEnum[keyof typeof GetAllProductsDirectionEnum];
/**
 * @export
 */
export const GetAllProductsStatusEnum = {
    Active: 'ACTIVE',
    Inactive: 'INACTIVE'
} as const;
export type GetAllProductsStatusEnum = typeof GetAllProductsStatusEnum[keyof typeof GetAllProductsStatusEnum];