// src/api/unit/api.ts
export interface UnitsApiGetAllUnitsRequest {
    /**
     * The starting index of the units.
     * @type {number}
     * @memberof UnitsApiGetAllUnits
     */
    readonly from?: number

    /**
     * The number of units to return.
     * @type {number}
     * @memberof UnitsApiGetAllUnits
     */
    readonly size?: number

    /**
     * Filter criteria for units.
     * @type {string}
     * @memberof UnitsApiGetAllUnits
     */
    readonly filter?: string
}