// src/api/fees/api.ts
export interface FeesApiGetAllFeesRequest {
    /**
     * The starting index of the fees.
     * @type {number}
     * @memberof FeesApiGetAllFees
     */
    readonly from?: number

    /**
     * The number of fees to return.
     * @type {number}
     * @memberof FeesApiGetAllFees
     */
    readonly size?: number

    /**
     * Filter criteria for fees.
     * @type {string}
     * @memberof FeesApiGetAllFees
     */
    readonly filter?: string

    /**
     * Field to order fees by.
     * @type {'createdAt'}
     * @memberof FeesApiGetAllFees
     */
    readonly orderBy?: GetAllFeesOrderByEnum

    /**
     * Sorting direction.
     * @type {'ASC' | 'DESC'}
     * @memberof FeesApiGetAllFees
     */
    readonly direction?: GetAllFeesDirectionEnum

    /**
     * Type of fee to filter by.
     * @type {'DISCOUNT' | 'TAX'}
     * @memberof FeesApiGetAllFees
     */
    readonly type?: GetAllFeesTypeEnum
}
// ...
export const GetAllFeesOrderByEnum = {
    CreatedAt: 'createdAt'
} as const;
export type GetAllFeesOrderByEnum = typeof GetAllFeesOrderByEnum[keyof typeof GetAllFeesOrderByEnum];
/**
 * @export
 */
export const GetAllFeesDirectionEnum = {
    Asc: 'ASC',
    Desc: 'DESC'
} as const;
export type GetAllFeesDirectionEnum = typeof GetAllFeesDirectionEnum[keyof typeof GetAllFeesDirectionEnum];
/**
 * @export
 */
export const GetAllFeesTypeEnum = {
    Discount: 'DISCOUNT',
    Tax: 'TAX'
} as const;
export type GetAllFeesTypeEnum = typeof GetAllFeesTypeEnum[keyof typeof GetAllFeesTypeEnum];